refactor(effects): split pulsing along register/cancel/tick seams
Splits the 1293-line cell/effects/pulsing.rs into a directory along its
natural lifecycle seams. Behavior-preserving pure code movement — no
logic or signature changes; re-export discipline keeps the public surface
identical (parent `mod pulsing;` + `pub use pulsing::{...}` unchanged).

Layout:
- pulsing/mod.rs          — module doc, shared consts
                            (MAX_CHANNEL_DURATION_SECS,
                            CHANNEL_INTERRUPT_DISTANCE), re-exports
- pulsing/register.rs     — register_active_effect
- pulsing/channel_cancel.rs — cancel_channels_from_attacker,
                            channel_interrupt_on_movement_tick,
                            cancel_channels_for_invoker_ability
- pulsing/tick.rs         — effect_pulse_tick, fire_pulse
- pulsing/tests.rs        — the 18 inline tests (shared make_mgr /
                            make_dot_effect harness kept whole)

Depth-shift fixes: former `super::EffectContext` / `super::dispatch_*`
references in the moved cancel/tick code now use absolute
`crate::cell::effects::` paths (one level deeper post-move). Shared
consts referenced via `super::` from submodules.

Part of #529.
